requirements.txt
flask
oslo.log
oslo.config
oslo.concurrency
python-keystoneclient
python-neutronclient
